+++ test/eca/features/chat_test.clj
@@ -17,7 +17,7 @@
         (with-redefs [llm-api/complete! (:api-mock mocks)
                       f.tools/call-tool! (:call-tool-mock mocks)
                       f.tools/approval (constantly :allow)]
-          (f.chat/prompt params (h/db*) (h/messenger) (h/config)))]
+          (f.chat/prompt params (h/db*) (h/messenger) (h/config) (h/metrics)))]
     (is (match? {:chat-id string? :status :prompting} resp))
     {:chat-id chat-id}))